+++ CHANGELOG.md
@@ -26,3 +26,4 @@
 
 - Adds Google Analytics events to user login, csv download, filter and table selection, custom search actions.
 - creates csv download filenames based on table parameters.
+- Adds analytics modal to explain our tracking